+++ apps/charterafrica/src/lib/data/common/index.js
@@ -3,18 +3,16 @@ import processPageArticles, {
   processPageEvents,
 } from "@/charterafrica/lib/data/common/processPageArticles";
 import processPageConsultation from "@/charterafrica/lib/data/common/processPageConsultation";
-import {
-  processPageDemocracyTools,
-  processPagePeople,
-  processPageOrganisations,
-} from "@/charterafrica/lib/data/common/processPageDemocracyTools";
 import processPageExplainers from "@/charterafrica/lib/data/common/processPageExplainers";
 import processPageIndex from "@/charterafrica/lib/data/common/processPageIndex";
 import processPageOpportunities, {
   processPageFellowships,
   processPageGrants,
 } from "@/charterafrica/lib/data/common/processPageOpportunities";
+import processPageOrganisations from "@/charterafrica/lib/data/common/processPageOrganisations";
+import processPagePeople from "@/charterafrica/lib/data/common/processPagePeople";
 import processPagePrivacyPolicy from "@/charterafrica/lib/data/common/processPagePrivacyPolicy";
+import processPageTools from "@/charterafrica/lib/data/common/processPageTools";
 import { getPageSeoFromMeta } from "@/charterafrica/lib/data/seo";
 
 export async function getGlobalProps({ locale, defaultLocale }, api) {
@@ -57,7 +55,7 @@ const processPageFunctionsMap = {
   index: processPageIndex,
   news: processPageArticles,
   research: processPageArticles,
-  tools: processPageDemocracyTools,
+  tools: processPageTools,
   "privacy-policy": processPagePrivacyPolicy,
   people: processPagePeople,
   organisations: processPageOrganisations,
