+++ apps/charterafrica/src/utils/translationConstants.js
@@ -1,0 +1,61 @@
+const labelsPerLocale = {
+  en: {
+    topic: "Topic A-Z",
+    "-topic": "Topic Z-A",
+    views: "Views Ascending",
+    "-views": "Views Descending",
+    stars: "Stars Ascending",
+    "-stars": "Stars Descending",
+    name: "Name",
+    goToRepo: "Go to Repo",
+    contribute: "Contribute",
+    tools: "Tools",
+    contributors: "Contributors",
+    active: "Active",
+    people: "People",
+    searchTools: "Search Tools",
+    organisations: "organisations",
+    searchPeople: "Search People",
+    searchOrganisations: "Search organisations",
+  },
+  fr: {
+    topic: "Sujet A-Z",
+    "-topic": "Sujet Z-A",
+    views: "Vues ascendantes",
+    "-views": "Vues descendant",
+    stars: "Étoiles ascendant",
+    "-stars": "Stars descendant",
+    name: "Nom",
+    goToRepo: "Aller au référentiel",
+    contribute: "Contribuer",
+    tools: "Outils",
+    contributors: "Contributeurs",
+    active: "Actif",
+    people: "Personnes",
+    searchTools: "Outils de recherche",
+    organisations: "Organisations",
+    searchPeople: "Recherche de personnes",
+    searchOrganisations: "Recherche d'organisations",
+  },
+  pt: {
+    topic: "Tópico A-Z",
+    "-topic": "Tópico Z-A",
+    views: "Visualizações ascendentes",
+    "-views": "Vistas descendentes",
+    stars: "Estrelas ascendendo",
+    "-stars": "Estrelas descendentes",
+    name: "Nome",
+    goToRepo: "Vá para o repositório",
+    contribute: "Contribuir",
+    tools: "Ferramentas",
+    contributors: "Colaboradores",
+    active: "Activo",
+    people: "Pessoas",
+    searchTools: "Ferramentas de busca",
+    organisations: "Organizaçãoes",
+    searchPeople: "Procurar pessoas",
+    searchOrganisations: "Procurar organizações",
+  },
+};
+
+export default labelsPerLocale;
